--- a/src/atopile/buildutil.py
+++ b/src/atopile/buildutil.py
@@ -6,8 +6,6 @@ from copy import deepcopy
 from pathlib import Path
 from typing import Callable, Optional
 
-from more_itertools import first
-
 from atopile import layout
 from atopile.config import config
 from atopile.errors import UserException, UserPickError
@@ -111,7 +109,7 @@ def build(app: Module) -> None:
     parameters = app.get_children(False, types=Parameter)
     if parameters:
         logger.info("Simplifying parameter graph")
-        solver.inspect_get_known_supersets(first(parameters), force_update=True)
+        solver.simplify(*parameters)
 
     # Pickers ------------------------------------------------------------------
     if config.build.keep_picked_parts:
@@ -339,6 +337,7 @@ def generate_module_map(app: Module, solver: Solver) -> None:
 def generate_variable_report(app: Module, solver: Solver) -> None:
     """Generate a report of all the variable values in the design."""
     # TODO: support other file formats
+    solver.update_superset_cache(app)
     export_parameters_to_file(
         app, solver, config.build.paths.output_base.with_suffix(".variables.md")
     )