--- a/src/atopile/errors.py
+++ b/src/atopile/errors.py
@@ -295,7 +295,9 @@ class UserPickError(UserException):
     def from_pick_error(cls, ex: "PickError") -> "UserPickError":
         from atopile.front_end import from_dsl
 
-        if origin_t := ex.module.try_get_trait(from_dsl):
+        modules = ex.module
+        # TODO not just first module
+        if modules and (origin_t := modules[0].try_get_trait(from_dsl)):
             origin = origin_t.src_ctx
         else:
             origin = None