--- a/src/faebryk/core/node.py
+++ b/src/faebryk/core/node.py
@@ -477,6 +477,8 @@ class Node(CNode):
         return out
 
     def _setup(self, *args, **kwargs) -> None:
+        assert not hasattr(self, "_setup_done")
+        self._setup_done = False
         # Construct Fields
         _, _ = self._setup_fields()
 
@@ -488,6 +490,7 @@ class Node(CNode):
                     if f_name in base.__dict__:
                         f = getattr(base, f_name)
                         f(self)
+        self._setup_done = True
 
     def __hash__(self):
         return id(self)
@@ -495,15 +498,15 @@ class Node(CNode):
     def __init__(self):
         super().__init__()
         CNode.transfer_ownership(self)
-        assert not hasattr(self, "_is_setup")
-        self._is_setup = True
+        assert not hasattr(self, "_called_init")
+        self._called_init = True
 
     def __preinit__(self, *args, **kwargs) -> None: ...
 
     def __postinit__(self, *args, **kwargs) -> None: ...
 
     def __post_init__(self, *args, **kwargs):
-        if not getattr(self, "_is_setup", False):
+        if not getattr(self, "_called_init", False):
             raise Exception(
                 "Node constructor hasn't been called."
                 "Did you forget to call super().__init__()?"
@@ -557,6 +560,10 @@ class Node(CNode):
     def builder(self, op: Callable[[Self], Any]) -> Self:
         op(self)
         return self
+
+    @property
+    def _is_setup(self) -> bool:
+        return getattr(self, "_setup_done", False)
 
     # printing -------------------------------------------------------------------------
 
@@ -779,7 +786,10 @@ class Node(CNode):
         return {n.get_name(): n for n in nodes}
 
     def __rich_repr__(self):
-        yield self.get_full_name()
+        if not self._is_setup:
+            yield f"{type(self)}(not init)"
+        else:
+            yield self.get_full_name()
 
     __rich_repr__.angular = True