… examples/ch2_3_minimal_led_explicit_parts.py
… examples/ch2_4_maximal_led.py
… examples/ch2_7_pcb_layout.py
… examples/ch2_8_mcu.py
… src/faebryk/core/solver/README.md
… src/faebryk/core/solver/solver.py
… src/faebryk/core/solver/symbolic/canonical.py
… src/faebryk/core/solver/symbolic/expression_groups.py
… src/faebryk/core/solver/symbolic/expression_wise.py
… src/faebryk/core/solver/symbolic/pure_literal.py
… src/faebryk/core/solver/symbolic/structural.py
… src/faebryk/core/solver/utils.py
… src/faebryk/library/LED.py
… src/faebryk/library/PoweredLED.py
… src/faebryk/library/ResistorVoltageDivider.py
… src/faebryk/library/TVS.py
… src/faebryk/library/can_attach_to_footprint_via_pinmap.py
… src/faebryk/library/has_construction_dependency.py
… src/faebryk/libs/logging.py
… src/faebryk/libs/picker/api/picker_lib.py
… src/faebryk/libs/picker/lcsc.py
… src/faebryk/libs/picker/localpick.py
… src/faebryk/libs/picker/picker.py
… src/faebryk/libs/sets/quantity_sets.py
… src/faebryk/libs/sets/sets.py
… src/faebryk/libs/test/times.py
… src/faebryk/libs/util.py
… test/core/performance/test_performance_pick.py
… test/core/solver/test_literal_folding.py
… test/core/solver/test_solver.py