# Accelerate picker

Baseline 3x3: 41.4s
pre-tb Current time: 7.88s
pre-tb Speedup: 5.25x
Current time: 13.55s
Speedup 3.05x

Baseline 4x4: 210s
pre-tb Current time: 27s
pre-tb Speedup: 7.7x
Current time: 43.71s
Speedup 4.8x